requirements.txt

pydantic
python-dotenv
python-multipart==0.0.9
requests==2.32.3

bcrypt==4.1.3
PyJWT==2.9.0
python-jose==3.3.0

supabase==2.6.0

mercadopago==2.2.3
google-auth==2.34.0

reportlab
fpdf2==2.7.9
python-barcode
pillow
qrcode[pil]
passlib[bcrypt]
bcrypt==4.1.3
passlib[bcrypt]
cryptography>=42.0.0
